+++ src/main/java/me/eting/server/model/models/age/AgeModelService.java
@@ -51,7 +51,7 @@ public class AgeModelService extends EtingModelService {
 
     @Override
     public int value(Incognito incognito) {
-        AgeModel model = ageModelRepository.findOne((long) incognito.getId());
+        AgeModel model = ageModelRepository.findOne(incognito.getId());
         if(model == null) return 0;
         return model.getValue();
     }
